move split objects under datasets
